Refactor wal with the new iterator
