+++ src/kvstore/wal/CMakeLists.txt
@@ -1,8 +1,8 @@
 nebula_add_library(
     wal_obj OBJECT
     InMemoryLogBuffer.cpp
-    FileBasedWalIterator.cpp
     FileBasedWal.cpp
+    WalFileIterator.cpp
 )
 
 nebula_add_subdirectory(test)
